Add function for adding text items to forms
