+++ create_form.py
@@ -2,7 +2,7 @@ from google.oauth2 import service_account
 from google.auth.transport.requests import AuthorizedSession
 from apiclient import discovery
 import json
-from create_question import dummy_round, create_gform_question, create_gform_round
+from create_gform_items import TEST_ROUND, create_gform_question, create_gform_round, create_gform_text_question
 
 # https://google-auth.readthedocs.io/en/master/user-guide.html
 
@@ -15,7 +15,7 @@ scoped_credentials = credentials.with_scopes(
 DISCOVERY_DOC = "https://forms.googleapis.com/$discovery/rest?version=v1"
 
 form_service = discovery.build('forms', 'v1', credentials=scoped_credentials,
-                               discoveryServiceUrl=DISCOVERY_DOC, 
+                               discoveryServiceUrl=DISCOVERY_DOC,
                                static_discovery=False)
 
 # End of Authorisation Code
@@ -27,7 +27,9 @@ form = {
 }
 
 # Creates the initial form
-result = form_service.forms().create(body=form).execute()
+QUIZ_FORM = form_service.forms().create(body=form).execute()
+QUIZ_FORM_ID = QUIZ_FORM["formId"]
+QUIZ_FORM_URL = QUIZ_FORM["responderUri"]
 
 
 # JSON to convert the form into a quiz & add description to a Form
@@ -42,6 +44,10 @@ def create_google_form(quiz_name, round_obj):
     """
     round_num = round_obj["round_num"]
     round_category = round_obj["question_data"][0]["category"]
+
+    form_questions = create_gform_round(round_obj)
+
+    # create_gform_text_question("Team Name")
 
     body = {
         "requests": [
@@ -59,31 +65,45 @@ def create_google_form(quiz_name, round_obj):
                 "updateFormInfo": {
                     "info": {
                         "title": quiz_name,
-                        "description": f"Round No. {round_num}: {round_category}"
+                        "description": f"Round {round_num}: {round_category}"
                     },
-                    "updateMask": "description"
+                    "updateMask": "*"
                 }
             },
-            create_gform_round(round_obj)
+            form_questions
         ]
     }
-
 
     return body
 
 
-update = create_google_form("My First Quiz", dummy_round)
-
-# print(update)
+update = create_google_form("My First Quiz", TEST_ROUND)
 
 # Updates the form
-question_setting = form_service.forms().batchUpdate(formId=result["formId"],
+question_setting = form_service.forms().batchUpdate(formId=QUIZ_FORM_ID,
                                                     body=update).execute()
 
+
+def add_item_to_gform(item):
+    """
+    Adds a Question to Google Form
+    """
+    body_text = {
+        "requests": [
+            item
+        ]
+    }
+    form_service.forms().batchUpdate(formId=QUIZ_FORM_ID, body=body_text).execute()
+
+
+team_name_q = create_gform_text_question("Team Name")
+
+add_item_to_gform(team_name_q)
+
 # Print the result to see it's now a quiz
-getresult = form_service.forms().get(formId=result["formId"]).execute()
+getresult = form_service.forms().get(formId=QUIZ_FORM["formId"]).execute()
 
-print(getresult)
+# print(getresult)
 
-form_id = getresult["formId"]
-print(f"Form ID = {form_id}")
+print(f"Form ID: {QUIZ_FORM_ID}")
+print(f"Form URL: {QUIZ_FORM_URL}")
